split deployment data types and protopuf types
