+++ src/routes/DeploymentDetail/ManifestEditor.js
@@ -2,7 +2,8 @@ import { useEffect, useState } from "react";
 import { Button, Box, Typography, LinearProgress, Tooltip, makeStyles } from "@material-ui/core";
 import { getDeploymentLocalData, saveDeploymentManifest } from "../../shared/utils/deploymentLocalDataUtils";
 import { TransactionMessageData } from "../../shared/utils/TransactionMessageData";
-import { NewDeploymentData, Manifest, sendManifestToProvider } from "../../shared/utils/deploymentUtils";
+import { sendManifestToProvider } from "../../shared/utils/deploymentUtils";
+import { deploymentData } from "../../shared/deploymentData";
 import { useWallet } from "../../context/WalletProvider";
 import { useTransactionModal } from "../../context/TransactionModal";
 import { useCertificate } from "../../context/CertificateProvider";
@@ -68,7 +69,7 @@ export function ManifestEditor({ deployment, leases, closeManifestEditor }) {
 
         const doc = yaml.load(yamlStr);
 
-        await NewDeploymentData(settings.apiEndpoint, doc, dseq, address);
+        await deploymentData.NewDeploymentData(settings.apiEndpoint, doc, dseq, address);
 
         setParsingError(null);
       } catch (err) {
@@ -115,8 +116,8 @@ export function ManifestEditor({ deployment, leases, closeManifestEditor }) {
 
   async function handleUpdateClick() {
     const doc = yaml.load(editedManifest);
-    const dd = await NewDeploymentData(settings.apiEndpoint, doc, parseInt(deployment.dseq), address); // TODO Flags
-    const mani = Manifest(doc);
+    const dd = await deploymentData.NewDeploymentData(settings.apiEndpoint, doc, parseInt(deployment.dseq), address); // TODO Flags
+    const mani = deploymentData.Manifest(doc);
 
     try {
       const message = TransactionMessageData.getUpdateDeploymentMsg(dd);
