+++ src/shared/deploymentData/index.js
@@ -1,0 +1,26 @@
+import * as v1beta1 from "./v1beta1";
+import * as v1beta2 from "./v1beta2";
+import { mainnetId, testnetId, edgenetId } from "../constants";
+export * from "./helpers";
+
+export let deploymentData;
+
+export function initDeploymentData() {
+  const selectedNetworkId = localStorage.getItem("selectedNetworkId");
+
+  switch (selectedNetworkId) {
+    case mainnetId:
+      deploymentData = v1beta1;
+      break;
+    case testnetId:
+      deploymentData = v1beta1;
+      break;
+    case edgenetId:
+      deploymentData = v1beta2;
+      break;
+
+    default:
+      deploymentData = v1beta1;
+      break;
+  }
+}
